Remove feltor/feltor.cu & feltor/window_params.js
The 3d code is so slow that it does not really make sense to watch it on
screen especially since afterwards all data is lost and we waited all
this time for nothing. Also now with the new adaptive Timesteppers we
don't need to use feltor.cu for testing out timesteps any more.