Fix config bug in cuda-aware support determination
also found a bug if mpi_inti2d and mpi_init3d are both used in one program
with cuda-aware mpi and 1 GPU. The segfault goes away if mpi_init2d and
mpi_init3d are called directly one after another. Not sure if we should
track this bug down.